chtrabata9

+++ src/Controller/TournamentsController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Security\EmailVerifier;
 use App\Entity\Tournaments;
 use App\Form\TournamentsType;
 use Doctrine\ORM\EntityManagerInterface;
@@ -10,10 +11,15 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Constraints\DateTime;
+use Symfony\Component\Mailer\MailerInterface;
 
 #[Route('/tournaments')]
 class TournamentsController extends AbstractController
 {
+    public function __construct(EmailVerifier $emailVerifier)
+    {
+        $this->emailVerifier = $emailVerifier;
+    }
     #[Route('/', name: 'app_tournaments_index', methods: ['GET'])]
     public function index(EntityManagerInterface $entityManager): Response
     {
@@ -35,19 +41,56 @@ class TournamentsController extends AbstractController
         $date = new \DateTime('now'); 
         $tournament->setCreateDate($date);
 
+
+       
+
         if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager = $form->getData();
             $entityManager->persist($tournament);
             $entityManager->flush();
 
+             // generate a signed url and email it to the user
+             $this->emailVerifier->sendEmailConfirmation('app_verify_email', $user,
+             (new TemplatedEmail())
+                 ->from(new Address('appgzone@gmail.com', 'Gzone App'))
+                 ->to($user->getEmail())
+                 ->subject('Please Confirm your Email')
+                 ->htmlTemplate('TournamentConfirmation/confirmation_TR.html.twig')
+         );
+         
+            
+
+            
             return $this->redirectToRoute('app_tournaments_index', [], Response::HTTP_SEE_OTHER);
         }
+
 
         return $this->renderForm('tournaments/new.html.twig', [
             'tournament' => $tournament,
             'form' => $form,
         ]);
     }
+ /**
+     * @Route("/verify/email", name="app_verify_email")
+     */
+    public function verifyUserEmail(Request $request): Response
+    {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
 
+        // validate email confirmation link, sets User::isVerified=true and persists
+        try {
+            $this->emailVerifier->handleEmailConfirmation($request, $this->getUser());
+        } catch (VerifyEmailExceptionInterface $exception) {
+            $this->addFlash('verify_email_error', $exception->getReason());
+
+            return $this->redirectToRoute('app_login');
+        }
+
+        // @TODO Change the redirect on success and handle or remove the flash message in your templates
+        $this->addFlash('success', 'Your email address has been verified.');
+
+        return $this->redirectToRoute('app_login');
+    }
     #[Route('/{id}', name: 'app_tournaments_show', methods: ['GET'])]
     public function show(Tournaments $tournament): Response
     {
@@ -80,6 +123,10 @@ class TournamentsController extends AbstractController
     #[Route('/{id}', name: 'app_tournaments_delete', methods: ['POST'])]
     public function delete(Request $request, Tournaments $tournament, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser()->getId() != $tournament->getAdmin()->getId()) {
+            return $this->redirectToRoute('app_tournaments_index', [], Response::HTTP_SEE_OTHER);
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$tournament->getId(), $request->request->get('_token'))) {
             $entityManager->remove($tournament);
             $entityManager->flush();
